@@ -4,13 +4,15 @@
   "private": true,
   "dependencies": {
     "@sentry/react": "^7.56.0",
+    "@supabase/auth-ui-react": "^0.4.6",
+    "@supabase/auth-ui-shared": "^0.1.8",
     "@supabase/supabase-js": "^2.38.3",
     "bootstrap": "^5.3.2",
     "dotenv": "^16.3.1",
     "lodash.debounce": "4.0.8",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-color": "2.19.3",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-file-utils": "1.1.1",
     "react-google-calendar-api": "^2.3.0",
     "react-scripts": "4.0.3",
