@@ -1,9 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import * as Sentry from '@sentry/react';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import 'bootstrap/dist/css/bootstrap.css';
+import { createRoot } from 'react-dom/client';
 
 if (process.env.NODE_ENV === 'production') {
   Sentry.init({
@@ -12,7 +12,9 @@ if (process.env.NODE_ENV === 'production') {
     tracesSampleRate: 1.0,
   });
 }
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App tab="home" />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
