@@ -1,97 +1,115 @@
-import React, { useEffect, useState } from 'react';
-import { StreamChat } from 'stream-chat';
-import { Chat, enTranslations, Streami18n } from 'stream-chat-react';
-import SignIn from './components/SignIn/SignIn.js';
-import Calendar from './components/Calendar';
-import Form from './components/Form.js'
-import { eventFormData, signFormData }from './assets/FormData.js'
+'use client';
+// import Link from 'next/link';
+import { useState, useEffect } from 'react'
+import { createClient } from '@supabase/supabase-js'
+import { Auth } from '@supabase/auth-ui-react'
+import MainChat from './components/MainChat/MainChat';
+import EventForm from './components/Forms/EventForm'
 
-import 'stream-chat-react/dist/css/index.css';
-import './App.css';
+const supabase = createClient(
+  'https://oodbxjicokcxmmclwojn.supabase.co', 
+  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im9vZGJ4amljb2tjeG1tY2x3b2puIiwicm9sZSI6ImFub24iLCJpYXQiOjE2ODU2NDQ3ODEsImV4cCI6MjAwMTIyMDc4MX0.RJHxPMDTBwx16ZElgEiOGNesdJac6340SKI5KAtih0k'
+)
 
-import { getRandomImage } from './assets';
-import { useChecklist } from './ChecklistTasks';
-import { ChannelContainer } from './components/ChannelContainer/ChannelContainer';
-import { ChannelListContainer } from './components/ChannelListContainer/ChannelListContainer';
-
-const urlParams = new URLSearchParams(window.location.search);
-const apiKey = urlParams.get('apikey') || process.env.REACT_APP_STREAM_KEY;
-const user = urlParams.get('user') || process.env.REACT_APP_USER_ID;
-const theme = urlParams.get('theme') || 'light';
-const userToken = urlParams.get('user_token') || process.env.REACT_APP_USER_TOKEN;
-const targetOrigin = urlParams.get('target_origin') || process.env.REACT_APP_TARGET_ORIGIN;
-
-const i18nInstance = new Streami18n({
-  language: 'en',
-  translationsForLanguage: {
-    ...enTranslations,
+const customTheme = {
+  default: {
+    colors: {
+      brand: 'hsl(153 60.0% 53.0%)',
+      brandAccent: 'hsl(154 54.8% 45.1%)',
+      brandButtonText: 'white',
+      // ..
+    },
   },
-});
+  dark: {
+    colors: {
+      brandButtonText: 'white',
+      defaultButtonBackground: '#2e2e2e',
+      defaultButtonBackgroundHover: '#3e3e3e',
+      //..
+    },
+  },
+  // You can also add more theme variations with different names.
+  evenDarker: {
+    colors: {
+      brandButtonText: 'white',
+      defaultButtonBackground: '#1e1e1e',
+      defaultButtonBackgroundHover: '#2e2e2e',
+      //..
+    },
+  },
+}
 
-const filters = [
-  { type: 'team', demo: 'team' },
-  { type: 'messaging', demo: 'team' },
-];
-const options = { state: true, watch: true, presence: true, limit: 3 };
-const sort = { last_message_at: -1, updated_at: -1 };
-
-const client = StreamChat.getInstance(apiKey, { enableInsights: true, enableWSFallback: true });
-client.connectUser({ id: user, name: user, image: getRandomImage() }, userToken);
-
-const App = () => {
-  const [createType, setCreateType] = useState('');
-  const [isCreating, setIsCreating] = useState(false);
-  const [isEditing, setIsEditing] = useState(false);
-
-  useChecklist(client, targetOrigin);
+export default function App() {
+  const [session, setSession] = useState(null)
 
   useEffect(() => {
-    const handleColorChange = (color) => {
-      const root = document.documentElement;
-      if (color.length && color.length === 7) {
-        root.style.setProperty('--primary-color', `${color}E6`);
-        root.style.setProperty('--primary-color-alpha', `${color}1A`);
+    supabase.auth.getSession().then(({ data: { session } }) => {
+      setSession(session)
+    })
+    
+    const {
+      data: { subscription },
+    } = supabase.auth.onAuthStateChange((_event, session) => {
+      setSession(session)
+    })
+
+    // (async function getUser() {
+    //   const {
+    //     data: { session },
+    //   } = await supabaseClient.auth.getSession()
+    //   console.log(data)
+    // })()
+
+    return () => subscription.unsubscribe()
+  }, [])
+
+  const getUserId = async () => {
+    let { data: profiles, error } = await supabase
+    .from('profiles')
+    .select('id')
+    return profiles
+  }
+
+  const insertUserId = async () => {
+    const { data, error } = await supabase
+      .from('profiles')
+      .insert([
+        { id: session.user.id },
+    ])
+  }
+
+  if(session) {
+    (async function get() {
+      const profiles = await getUserId()
+      console.log(profiles)
+      if (profiles.length === 0) {
+        insertUserId()
+      } else {
+        return
       }
-    };
+    })()
+  }
 
-    window.addEventListener('message', (event) => handleColorChange(event.data));
-    return () => {
-      client.disconnectUser();
-      window.removeEventListener('message', (event) => handleColorChange(event.data));
-    };
-  }, []);
-
-  return (
-    <>
-      <div className='app__wrapper'>
-        <Chat {...{ client, i18nInstance }} theme={`team ${theme}`}>
-          <ChannelListContainer
-            {...{
-              isCreating,
-              filters,
-              options,
-              setCreateType,
-              setIsCreating,
-              setIsEditing,
-              sort,
-            }}
-          />
-          <ChannelContainer
-            {...{
-              createType,
-              isCreating,
-              isEditing,
-              setIsCreating,
-              setIsEditing,
-            }}
-          />
-        </Chat>
+  if (!session) {
+    return (
+      <div class="container w-50">
+        <Auth 
+          supabaseClient={supabase}
+          theme="default" // can also be "dark" or "evenDarker"
+          appearance={{ 
+            theme: customTheme, 
+            className: {
+              input: 'input-h',
+              button: 'my-awesome-button',
+              //..
+            },
+          }} 
+        />
       </div>
-      <Calendar />
-      <Form formData={signFormData} />
-      <Form formData={eventFormData} />
-    </>
-  );
-};
-
-export default App;
+    )
+  } else {
+    return (
+      <EventForm />
+    )
+  }
+}
